Move security logic out of msg.
