+++ contracts/nav/src/util/mod.rs
@@ -1,3 +1,4 @@
 pub mod action;
+pub mod security;
 pub mod state;
 pub mod validate;
